refactor: breaking out mutation record callbacks into separate files
