+++ src/callbacks/ContentEditableChange.js
@@ -1,0 +1,8 @@
+const isContentEditableChildList = (mutation) =>
+  mutation.type === 'childList' &&
+  mutation.target.hasAttribute('contentEditable');
+
+export const handleContentEditableChange = (action) => (mutations) =>
+  mutations
+    .filter(isContentEditableChildList)
+    .forEach((mutation) => action(mutation.target));
