+++ src/callbacks/ContentEditableInsert.js
